refactor(event) : 공통 로직 추상 클래스로 분리, model/event 패키지 추가,이동
