+++ src/main/java/christmas/model/event/PresentEventStrategy.java
@@ -1,14 +1,12 @@
-package christmas.model;
+package christmas.model.event;
 
-import static christmas.util.Constants.DASH;
 import static christmas.util.Constants.ZERO;
 
-import christmas.util.Parser;
 import christmas.util.enums.EventType;
 import java.util.LinkedHashMap;
 import java.util.Map;
 
-public class PresentEventStrategy implements EventStrategy {
+public class PresentEventStrategy extends EventDetailsHandler implements EventStrategy {
     private static final int PRESENT_DISCOUNT = 25000;
     private static final int PRESENT_RANGE = 120000;
 
@@ -51,24 +49,11 @@ public class PresentEventStrategy implements EventStrategy {
         return details;
     }
 
-    private void putNone(final EventType eventType, final Map<String, String> details) {
-        details.put(eventType.getNone(), eventType.getNone());
-    }
-
-    private void putDetails(final EventType eventType, final Integer discount,
-                                   final Map<String, String> details) {
-        details.put(eventType.getDescription(), DASH + Parser.decimalFormatter(discount));
-    }
-
     public String getPresent() {
         if (isPresentRange()) {
             return getEventType().getPresentDescription();
         }
         return getEventType().getNone();
-    }
-
-    private boolean isValid(final Integer discount) {
-        return discount > ZERO;
     }
 
     private boolean isPresentRange() {
