+++ src/main/java/christmas/model/event/WeekendEventStrategy.java
@@ -1,15 +1,14 @@
-package christmas.model;
+package christmas.model.event;
 
 import static christmas.util.Constants.*;
 
-import christmas.util.Parser;
 import christmas.util.enums.EventType;
 import java.time.DayOfWeek;
 import java.time.LocalDate;
 import java.util.LinkedHashMap;
 import java.util.Map;
 
-public class WeekendEventStrategy implements EventStrategy {
+public class WeekendEventStrategy extends EventDetailsHandler implements EventStrategy {
     private static final int WEEK_END_DISCOUNT = 2023;
 
     private final LocalDate date;
@@ -51,19 +50,6 @@ public class WeekendEventStrategy implements EventStrategy {
         }
         putNone(eventType, details);
         return details;
-    }
-
-    private boolean isValid(final Integer discount) {
-        return discount > ZERO;
-    }
-
-    private void putNone(final EventType eventType, final Map<String, String> details) {
-        details.put(eventType.getNone(), eventType.getNone());
-    }
-
-    private void putDetails(final EventType eventType, final Integer discount,
-                            final Map<String, String> details) {
-        details.put(eventType.getDescription(), DASH + Parser.decimalFormatter(discount));
     }
 
     private int calculateDiscount() {
