+++ src/test/java/christmas/model/DdayEventStrategyTest.java
@@ -2,6 +2,8 @@ package christmas.model;
 
 import static org.assertj.core.api.Assertions.*;
 
+import christmas.model.event.DdayEventStrategy;
+import christmas.model.event.EventStrategy;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
