+++ src/test/java/christmas/model/WeekdayEventStrategyTest.java
@@ -2,6 +2,7 @@ package christmas.model;
 
 import static org.assertj.core.api.Assertions.*;
 
+import christmas.model.event.WeekdayEventStrategy;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
