+++ src/test/java/christmas/model/WeekendEventStrategyTest.java
@@ -2,6 +2,7 @@ package christmas.model;
 
 import static org.assertj.core.api.Assertions.*;
 
+import christmas.model.event.WeekendEventStrategy;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
